Allow MPIX::Execution use short int

+++ src/lib/framework/MACE/Extension/MPIX/Execution/Scheduler.inl
@@ -1,7 +1,7 @@
 namespace MACE::inline Extension::MPIX::inline Execution {
 
 template<std::integral T>
-    requires(Concept::MPIPredefined<T> and sizeof(T) >= sizeof(int))
+    requires(Concept::MPIPredefined<T> and sizeof(T) >= sizeof(short))
 auto Scheduler<T>::Reset() -> void {
     fExecutingTask = fTask.first;
     fNLocalExecutedTask = 0;
